+++ docs/change-log.md
@@ -3,6 +3,7 @@
 ---
 ## 4 Aug 2016: V 1.7.2 (by FORTH-ICS)
 * Fixed a bug that was raised with multiple input files ([issue #49](https://github.com/isl/x3ml/issues/49))
+* Fixed a bug with NullPointerException that is thrown when the X3ML mappings file is not valid with respect to the X3ML schema ([issue #63](https://github.com/isl/x3ml/issues/63))
 * Changed the hard-coded way of reporting error messages with a configurable that exploits loggers ([issue #48](https://github.com/isl/x3ml/issues/48))
 * NEW functionality for using an entire folder (contain XML files) as input  ([issue #47](https://github.com/isl/x3ml/issues/47))
 * NEW functionality for searching recursively for XML files in an input folder ([issue #62](https://github.com/isl/x3ml/issues/62))